requirements.txt
argparse~=1.4.0
packaging~=24.0
numpy~=1.26.4
scipy~=1.13.0
diffusers~=0.27.2
accelerate~=0.29.2
torchvision~=0.18.0
huggingface_hub~=0.22.2
torchmetrics~=1.3.2
tqdm~=4.66.2
batchgenerators~=0.25
nibabel~=5.2.1
pillow~=10.3.0
